refactor: Вынесены отдельно компоненты попапов редактирования профиля и аватара, перенесены запросы по к серверу в App, создан контекст текущего пользователя добавлен функционал удаления и лайка карточки+функционал обновления инфо пользователя и аватара через управляемые компоненты и реф

+++ src/utils/Api.js
@@ -49,17 +49,18 @@ class Api {
     });
   }
   // лайки (Афродита одобрит)
-  addLike(cardId) {
-    return this._request(`${this._url}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    });
-  }
-  removeLike(cardId) {
-    return this._request(`${this._url}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    });
+  changeLikeCardStatus(cardId, isLiked) {
+    if (!isLiked) {
+      return this._request(`${this._url}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      });
+    } else {
+      return this._request(`${this._url}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      });
+    }
   }
 
   editUserInfo(data) {
